+++ spec/callable_tree/node/internal_spec.rb
@@ -14,6 +14,19 @@ RSpec.describe CallableTree::Node::Internal do
     it { is_expected.not_to be node.children }
   end
 
+  describe '#children!' do
+    subject { node.children! }
+
+    let(:node) do
+      CallableTree::Node::Root.new.append(
+        IdLeaf.new(:a),
+        IdLeaf.new(:b)
+      )
+    end
+
+    it { is_expected.to be node.children! }
+  end
+
   describe '#[]' do
     subject { node[nth] }